+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Development ()
 
+* **Bugfix**: Fix menu styling
+
 ## 0.13.0 (08.10.2020)
 
 * **New**: Allow marking paragraphs as extracts